+++ topics/Topic-2-Outage-Analysis.py
@@ -21,6 +21,8 @@ st.write("---")
 
 ## Main Content
 
+st.warning('This dashboard is a work in progress. It is intended for educational purposes only. If you have any feedback \
+           or suggestions, please feel free to open an issue on the [GitHub](https://github.com/ZhipengHe/NEM-Dashboard/issues).', icon="⚠️")
 
 st.write("---")
 
@@ -35,10 +37,10 @@ st.info('''
 st.markdown('''
             ##### Available Datasets
 
-            The data used in the above analysis is sourced from the Australian Energy Market Operator (AEMO) and is publicly available \
+            These tables are publicly available \
             on the [NEMWEB](https://aemo.com.au/energy-systems/electricity/national-electricity-market-nem/data-nem/market-data-nemweb). \
             Please refer to the [MMS Data Model](https://nemweb.com.au/Reports/Current/MMSDataModelReport/Electricity/Electricity%20Data%20Model%20Report.htm) \
-            for detailed information on the data structure and definitions. Here lists the relevant table names and features from MMS Data Model used in above example:
+            for detailed information on the data structure and definitions. 
             - [`NETWORK.NETWORK_OUTAGEDETAIL`](https://nemweb.com.au/Reports/Current/MMSDataModelReport/Electricity/Electricity%20Data%20Model%20Report_files/Elec72.htm#15): \
             This table contains detailed information about network outages, including the start and end times, duration, affected substations, \
             equipment types, and outage causes.